pyproject.toml
[project]
name = "managesg-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.0,<2.0.0",
    "google-generativeai>=0.7.0",
    "gspread==6.1.2",
    "langchain==0.3.27",
    "langchain-community==0.3.29",
    "langchain-core==0.3.75",
    "langchain-google-genai>=2.0.0",
    "langchain-mcp-adapters==0.0.7",
    "langchain-mongodb==0.7.0",
    "langchain-openai==0.3.32",
    "langgraph==0.6.6",
    "openai==1.102.0",
    "pdf2image==1.17.0",
    "pillow>=10.0.0",
    "pinecone>=8.0.0",
    "pinecone-client>=6.0.0",
    "pycryptodome>=3.15.0",
    "pydantic==2.11.7",
    "pymongo==4.6.3",
    "pymupdf>=1.26.7",
    "pymupdf4llm==0.0.27",
    "pytesseract==0.3.13",
    "python-docx==1.2.0",
    "python-dotenv==1.1.1",
    "python-pptx==1.0.2",
    "streamlit>=1.53.0",
    "tiktoken>=0.12.0",
]
